+++ src/smart-components/access-management/UsersTable.tsx
@@ -1,19 +1,21 @@
-import React, { useEffect, useCallback } from 'react';
+import React, { useEffect, useCallback, useState, Fragment, useMemo } from 'react';
 import { useSelector, useDispatch } from 'react-redux';
 import { useDataViewSelection, useDataViewPagination } from '@patternfly/react-data-view/dist/dynamic/Hooks';
 import { BulkSelect, BulkSelectValue } from '@patternfly/react-component-groups/dist/dynamic/BulkSelect';
 import { DataView } from '@patternfly/react-data-view/dist/dynamic/DataView';
 import { DataViewToolbar } from '@patternfly/react-data-view/dist/dynamic/DataViewToolbar';
 import { DataViewTable } from '@patternfly/react-data-view/dist/dynamic/DataViewTable';
-import { Button, Pagination } from '@patternfly/react-core';
+import { Button, Pagination, ButtonVariant } from '@patternfly/react-core';
 import { ActionsColumn } from '@patternfly/react-table';
 import { fetchUsers } from '../../redux/actions/user-actions';
 import { mappedProps } from '../../helpers/shared/helpers';
 import { RBACStore } from '../../redux/store';
-import { UserProps } from '../user/user-table-helpers';
-import { useSearchParams } from 'react-router-dom';
+import { User } from '../../redux/reducers/user-reducer';
 import { useIntl } from 'react-intl';
 import messages from '../../Messages';
+import { useSearchParams } from 'react-router-dom';
+import { WarningModal } from '@patternfly/react-component-groups';
+import { UserProps } from '../user/user-table-helpers';
 
 const COLUMNS: string[] = ['Username', 'Email', 'First name', 'Last name', 'Status', 'Org admin'];
 
@@ -32,14 +34,14 @@ interface UsersTableProps {
 }
 
 const UsersTable: React.FunctionComponent<UsersTableProps> = ({ onAddUserClick }) => {
+  const [isDeleteModalOpen, setIsDeleteModalOpen] = useState(false);
+  const [currentUser, setCurrentUser] = useState<User | undefined>();
   const dispatch = useDispatch();
   const intl = useIntl();
 
-  const rowActions = (user: UserProps) => {
-    return [
-      { title: intl.formatMessage(messages['usersAndUserGroupsAddToGroup']), onClick: () => onAddUserClick([user]) },
-      { title: intl.formatMessage(messages['usersAndUserGroupsRemoveFromGroup']), onClick: () => console.log('REMOVE FROM USER GROUP') },
-    ];
+  const handleModalToggle = (_event: KeyboardEvent | React.MouseEvent, user: User) => {
+    setCurrentUser(user);
+    setIsDeleteModalOpen(!isDeleteModalOpen);
   };
 
   const { users, totalCount } = useSelector((state: RBACStore) => ({
@@ -81,22 +83,38 @@ const UsersTable: React.FunctionComponent<UsersTableProps> = ({ onAddUserClick }
     }
   };
 
-  const rows = users.map((user: UserProps) => ({
-    id: user.username,
-    is_active: user.is_active,
-    row: [
-      user.username,
-      user.email,
-      user.first_name,
-      user.last_name,
-      user.is_active ? intl.formatMessage(messages['usersAndUserGroupsActive']) : intl.formatMessage(messages['usersAndUserGroupsInactive']),
-      user.is_org_admin ? intl.formatMessage(messages['usersAndUserGroupsYes']) : intl.formatMessage(messages['usersAndUserGroupsNo']),
-      {
-        cell: <ActionsColumn items={rowActions(user)} isDisabled={!user.is_active} />,
-        props: { isActionCell: true },
-      },
-    ],
-  }));
+  const rows = useMemo(() => {
+    return users.map((user: UserProps) => ({
+      id: user.username,
+      is_active: user.is_active,
+      row: [
+        user.username,
+        user.email,
+        user.first_name,
+        user.last_name,
+        user.is_active ? intl.formatMessage(messages['usersAndUserGroupsActive']) : intl.formatMessage(messages['usersAndUserGroupsInactive']),
+        user.is_org_admin ? intl.formatMessage(messages['usersAndUserGroupsYes']) : intl.formatMessage(messages['usersAndUserGroupsNo']),
+        {
+          cell: (
+            <ActionsColumn
+              items={[
+                {
+                  title: intl.formatMessage(messages['usersAndUserGroupsAddToGroup']),
+                  onClick: () => onAddUserClick([user]),
+                },
+                {
+                  title: intl.formatMessage(messages['usersAndUserGroupsRemoveFromGroup']),
+                  onClick: (event: KeyboardEvent | React.MouseEvent, rowId: number, rowData: any) => handleModalToggle(event, rowData),
+                },
+              ]}
+              rowData={user}
+            />
+          ),
+          props: { isActionCell: true },
+        },
+      ],
+    }));
+  }, [users, intl, onAddUserClick, handleModalToggle]);
 
   const pageSelected = rows.length > 0 && rows.every(isSelected);
   const pagePartiallySelected = !pageSelected && rows.some(isSelected);
@@ -113,7 +131,24 @@ const UsersTable: React.FunctionComponent<UsersTableProps> = ({ onAddUserClick }
   );
 
   return (
-    <>
+    <Fragment>
+      {isDeleteModalOpen && (
+        <WarningModal
+          ouiaId={`${OUIA_ID}-remove-user-modal`}
+          isOpen={isDeleteModalOpen}
+          title={intl.formatMessage(messages.deleteUserModalTitle)}
+          confirmButtonLabel={intl.formatMessage(messages.remove)}
+          confirmButtonVariant={ButtonVariant.danger}
+          onClose={() => setIsDeleteModalOpen(false)}
+          onConfirm={() => {
+            console.log(`Deleting ${currentUser?.username} from user groups`);
+            //add delete user api call here when v2 is ready
+            setIsDeleteModalOpen(false);
+          }}
+        >
+          {`${currentUser?.username} ${intl.formatMessage(messages.deleteUserModalBody)}`}
+        </WarningModal>
+      )}
       <DataView ouiaId={OUIA_ID} selection={{ ...selection, isSelectDisabled: (row) => !row.is_active }}>
         <DataViewToolbar
           ouiaId={`${OUIA_ID}-header-toolbar`}
@@ -145,7 +180,7 @@ const UsersTable: React.FunctionComponent<UsersTableProps> = ({ onAddUserClick }
         <DataViewTable variant="compact" aria-label="Users Table" ouiaId={`${OUIA_ID}-table`} columns={COLUMNS} rows={rows} />
         <DataViewToolbar ouiaId={`${OUIA_ID}-footer-toolbar`} pagination={paginationComponent} />
       </DataView>
-    </>
+    </Fragment>
   );
 };