+++ src/smart-components/access-management/users-and-user-groups.tsx
@@ -8,6 +8,8 @@ import UserGroupsTable from './UserGroupsTable';
 import { useLocation, useNavigate } from 'react-router-dom';
 import AddUserGroupModal from './AddUserGroupModal';
 
+const TAB_NAMES = ['users', 'user-groups'];
+
 const UsersAndUserGroups: React.FunctionComponent = () => {
   const intl = useIntl();
   const [activeTabKey, setActiveTabKey] = React.useState<number>(0);
@@ -20,7 +22,7 @@ const UsersAndUserGroups: React.FunctionComponent = () => {
   const navigate = useNavigate();
   const location = useLocation();
 
-  const updateURL = (tabKey: number) => {
+  const updateURL = (tabKey: string) => {
     const params = new URLSearchParams(location.search);
     params.set('activeTab', tabKey.toString());
     navigate({ search: params.toString() });
@@ -29,7 +31,7 @@ const UsersAndUserGroups: React.FunctionComponent = () => {
   const handleTabSelect = (_: React.MouseEvent<HTMLElement, MouseEvent>, key: string | number) => {
     const activeTab = Number(key);
     setActiveTabKey(activeTab);
-    updateURL(activeTab);
+    updateURL(TAB_NAMES[activeTab]);
   };
 
   const handleOpenAddUserModal = (selected: any[]) => {
@@ -42,7 +44,7 @@ const UsersAndUserGroups: React.FunctionComponent = () => {
   useEffect(() => {
     const params = new URLSearchParams(location.search);
     const tabKey = params.get('activeTab');
-    tabKey && setActiveTabKey(Number(tabKey));
+    tabKey && setActiveTabKey(Number(TAB_NAMES.findIndex((val) => val === tabKey)));
   }, [location.search]);
 
   return (