+++ CHANGELOG.md
@@ -1,3 +1,63 @@
+# [1.4.0](https://github.com/RedHatInsights/insights-rbac-ui/compare/v1.3.0...v1.4.0) (2024-10-24)
+
+
+### Features
+
+* **workspaces:** enable overview if feature flag enabled ([#1683](https://github.com/RedHatInsights/insights-rbac-ui/issues/1683)) ([59fa607](https://github.com/RedHatInsights/insights-rbac-ui/commit/59fa60744a0fe99304f4cda2795c0a2120036c49))
+
+# [1.3.0](https://github.com/RedHatInsights/insights-rbac-ui/compare/v1.2.0...v1.3.0) (2024-10-24)
+
+
+### Bug Fixes
+
+* memoize users row data ([60deb9b](https://github.com/RedHatInsights/insights-rbac-ui/commit/60deb9b76285d3b6899ebe59d8d849f92ae9a384))
+
+
+### Features
+
+* **userstable:** add delete user modal ([814f7f1](https://github.com/RedHatInsights/insights-rbac-ui/commit/814f7f161c513f7fe3bfd95c4909f30e10c18f30))
+
+# [1.2.0](https://github.com/RedHatInsights/insights-rbac-ui/compare/v1.1.0...v1.2.0) (2024-10-23)
+
+
+### Features
+
+* **konflux:** enablet unit tests on Konflux ([#1656](https://github.com/RedHatInsights/insights-rbac-ui/issues/1656)) ([527b4d8](https://github.com/RedHatInsights/insights-rbac-ui/commit/527b4d89273077820459004049fbd16b79e42d2b))
+
+# [1.1.0](https://github.com/RedHatInsights/insights-rbac-ui/compare/v1.0.1...v1.1.0) (2024-10-22)
+
+
+### Features
+
+* show WS enable alert only if user can perform this action ([#1681](https://github.com/RedHatInsights/insights-rbac-ui/issues/1681)) ([b89b2d3](https://github.com/RedHatInsights/insights-rbac-ui/commit/b89b2d3467f34996a36e5bd1052688fa4f021648))
+
+## [1.0.1](https://github.com/RedHatInsights/insights-rbac-ui/compare/v1.0.0...v1.0.1) (2024-10-22)
+
+
+### Bug Fixes
+
+* upgrade utility classes to PF5 ([750a18a](https://github.com/RedHatInsights/insights-rbac-ui/commit/750a18a60cdeacbee029b70ff1e7920f7b83e8f3))
+
+# 1.0.0 (2024-10-18)
+
+
+### Bug Fixes
+
+* remove mapStateToProps and mapDispatchToProps ([d9179d6](https://github.com/epwinchell/insights-rbac-ui/commit/d9179d636346c3cbb00401820254ed00ff88de41))
+* remove withRouter decorators ([5b17ed2](https://github.com/epwinchell/insights-rbac-ui/commit/5b17ed29b8a56abd124257d11d62b033c930e45c))
+* replace useRouteMatch with useParams ([d23a2a2](https://github.com/epwinchell/insights-rbac-ui/commit/d23a2a25469e821d41ac14bb70b396f43402db53))
+
+
+### Features
+
+* **rbac:** add access-requests ([#754](https://github.com/epwinchell/insights-rbac-ui/issues/754)) ([7fa22b0](https://github.com/epwinchell/insights-rbac-ui/commit/7fa22b0bd7ee61f09e266e09ac2691b1f011fd2f))
+
+
+### Reverts
+
+* Revert "Do not send username when creating service account ([#1559](https://github.com/epwinchell/insights-rbac-ui/issues/1559))" ([#1560](https://github.com/epwinchell/insights-rbac-ui/issues/1560)) ([a310c88](https://github.com/epwinchell/insights-rbac-ui/commit/a310c885ab44d063e378393b75f161aa04670704))
+* Revert "Revert "Do not send username when creating service account ([#1559](https://github.com/epwinchell/insights-rbac-ui/issues/1559))" ([#1560](https://github.com/epwinchell/insights-rbac-ui/issues/1560))" ([#1561](https://github.com/epwinchell/insights-rbac-ui/issues/1561)) ([38dc92f](https://github.com/epwinchell/insights-rbac-ui/commit/38dc92fea5c5cd4db8a3dd1cbfe90cce572656a1))
+
 # 1.0.0 (2024-10-17)